Fix fatal error in PHP 8 due invalid callback
The callback of add_settings_section() is set to a function that doesn't exist for mmj_pluginPage.

This was a warning in PHP 7 but is a fatal error in PHP 8.

We are creating the mmj_options_section_callback() function to prevent this issue. The function is empty because we don't need to print anything.. but the function must exist.